fix: header stats
-Fix bug where header stats displays 0 for all stats once lessons are completed and score is updated to db. -Refactor weekly stats update so that it doesn't run more times than intended by placing a condition that it doesn't update a second time until the 'switch weekly date' arrow buttons are clicked. Create utility functions to handle start and end date calculations. -Remove redundant weekly start/end date calculation hook.
