+++ client/src/components/forms/homepg/DifficultySettings.tsx
@@ -1,4 +1,11 @@
-import { useState, useRef, useContext, useEffect, useMemo, useLayoutEffect } from "react";
+import {
+  useState,
+  useRef,
+  useContext,
+  useEffect,
+  useMemo,
+  useLayoutEffect,
+} from "react";
 import { Fragment } from "react";
 import { v4 as uuidv4 } from "uuid";
 import { MenuContext } from "../../../providers/MenuProvider";
@@ -58,7 +65,6 @@ function SettingName({ inputRef }: SettingNameProps) {
         </span>
       );
     } else if (handleExistingName()) {
-      console.log("runs");
       return (
         <span className="pt-2 text-sm text-red-400">
           **Setting name already exists**
