+++ client/src/utils/calculations/CalculateStartEndDates.ts
@@ -1,0 +1,15 @@
+export default function CalculateStartEndDates({ numWeeksBeforeToday }) {
+  const daysInAWeek = 7;
+  const oneDayInSeconds = 86400000;
+
+  const generatedStartDate = new Date(
+    new Date().valueOf() - oneDayInSeconds * numWeeksBeforeToday * daysInAWeek,
+  ); //Todays Date
+
+  const generatedEndDate = new Date(
+    new Date().valueOf() -
+      oneDayInSeconds * (numWeeksBeforeToday * daysInAWeek + daysInAWeek),
+  ); //One week ago
+
+  return { generatedStartDate, generatedEndDate };
+}
